requirements.txt
coloredlogs
Pillow
polars
pandas
torch
torchvision
pyyaml